Avoid allocating IP node addrs when not required
When 9fa9ce1b was added, one of the changes that it introduced was to
allocate an IP address per node in the network for each network that the
node was attached to.  While this was a safe approach from a code
standpoint, it causes problems when upgrading to CE from older versions
because users that have networks allocated at near capacity suddenly
find their existing networks failing due to a lack of IP addresses.

This patch adjusts the behavior of swarmkit so that it only allocates
these extra IP addresses if the node is known to actually need it.  This
occurs when the network being attached to is marked as Ingress or when
the node is running Windows.

Signed-off-by: Chris Telfer <ctelfer@docker.com>